Code for staging

+++ Dev/Server/HtmlDeleter.py
@@ -1,10 +1,13 @@
 import os
+import os.path
 import shutil
 def deleteHTML(theJson, htmlDir):
     id = theJson.get("id")
     homeDir = htmlDir + "/" + id
-    shutil.rmtree(homeDir)
+    if os.path.exists(homeDir):
+        shutil.rmtree(homeDir)
 
 def deleteHTMLById(id, htmlDir):
     homeDir = htmlDir + "/" + id
-    shutil.rmtree(homeDir)
+    if os.path.exists(homeDir):
+        shutil.rmtree(homeDir)
